+++ lib/Controller/EditorApiController.php
@@ -1,7 +1,7 @@
 <?php
 /**
  *
- * (c) Copyright Ascensio System SIA 2025
+ * (c) Copyright Ascensio System SIA 2026
  *
  * This program is a free software product.
  * You can redistribute it and/or modify it under the terms of the GNU Affero General Public License
@@ -935,7 +935,7 @@ class EditorApiController extends OCSController {
                 $tags = $watermarkSettings["linkTagsList"];
                 $fileTags = Server::get(ISystemTagObjectMapper::class)->getTagIdsForObjects([$fileId], "files")[$fileId];
                 foreach ($fileTags as $tagId) {
-                    if (in_array($tagId, $tags, true)) {
+                    if (in_array($tagId, $tags)) {
                         return $watermarkText;
                     }
                 }